requirements.txt
streamlit==1.46.0 
pandas==2.1.4
openpyxl==3.1.2
streamlit-aggrid==0.3.4
google-generativeai==0.3.2
plotly==5.17.0
markdown==3.5.2
matplotlib==3.8.2 